Failed: Reading file in FE AFTER BE update
Failed attempts (some commented, some deleted) to read the file in the frontend but after the BE updates it.
"require" method seems to always immediately read the file.
Did not successfully solve this matter using async programming
All of the former was mostly done at around 1 PM on August 1st, 2021